stock-service/src/main/java/com/emazon/stock_service/domain/api/usecase/CategoryUseCase.java
package com.emazon.stock_service.domain.api.usecase;

import com.emazon.stock_service.domain.model.Category;
import com.emazon.stock_service.domain.spi.ICategoryPersistencePort;
import com.emazon.stock_service.domain.api.ICategoryServicePort;
import org.springframework.web.bind.annotation.*;

@RestController
@RequestMapping("/api/categories")

public class CategoryUseCase implements ICategoryServicePort {

    private final ICategoryPersistencePort categoryPersistencePort;

    public CategoryUseCase(ICategoryPersistencePort categoryPersistencePort) {
        this.categoryPersistencePort = categoryPersistencePort;
    }

    @Override
    public void saveCategory(Category category) {

        categoryPersistencePort.saveCategory(category);
    }

}






